Remove old files_opds 3rd-party app integration
The metadata service is currently unused and it's guarded behind a condition that checks if the files_opfs app exists.
That app doesn't work and it hasn't been updated for 7 years so let's get rid of it:
https://github.com/Yetangitu/owncloud-apps/tree/master/files_opds

Note: it doesn't do anything useful outside of files_opds, unlike its name suggests it's not metadata in general but only files_opds-provided metadata.