Fix `Psr\Log\AbstractLogger` error
Nextcloud broke things again in an update without mentioning it in the changelog. A ton of apps are impacted (not just epubviewer).

This commit uses a trick to prevent epubviewer's dependencies from installing `Psr\Log` and instead use the one diretly shipped in the dependencies of Nextcloud 31.

Fix https://github.com/devnoname120/epubviewer/discussions/29#discussioncomment-12334932

Related:
- https://github.com/nextcloud/mail/pull/10644
- https://github.com/ChristophWurst/nextcloud_sentry/pull/701
- https://github.com/nextcloud/mail/pull/10706
- https://github.com/nextcloud/mail/issues/10658
- ihttps://github.com/nextcloud/mail/pull/10708
- And many more…